Extract openscenario_preprocessor
