+++ test_runner/scenario_test_runner/package.xml
@@ -14,6 +14,7 @@
 
   <exec_depend>simple_sensor_simulator</exec_depend>
   <exec_depend>openscenario_interpreter</exec_depend>
+  <exec_depend>openscenario_preprocessor</exec_depend>
   <exec_depend>openscenario_utility</exec_depend>
 
   <!-- AutowareAuto launch -->
